feat: configure os
First, in order to better organize screenshots by section, moved the previous screenshots to new subdirectory and created a new subdirectory for the new ones.

Also, installing the OS may be more straight-forward and obvious to most users, so I moved it to a second file so that it doesn't clutter the more important steps.

Finally, this adds instructions for the initial setup of the OS short of actually installing RetroPie itself.